+++ src/plone/app/fhirfield/deserializer.py
@@ -1,9 +1,11 @@
 # _*_ coding: utf-8 _*_
 import six
 from plone.app.fhirfield.interfaces import IFhirResource
+from plone.app.fhirfield.interfaces import IFhirResourceValue
 from plone.dexterity.interfaces import IDexterityContent
 from plone.restapi.deserializer.dxfields import DefaultFieldDeserializer
 from plone.restapi.interfaces import IFieldDeserializer
+from pydantic import BaseModel
 from zope.component import adapter
 from zope.interface import implementer
 from zope.publisher.interfaces.browser import IBrowserRequest
@@ -21,6 +23,10 @@ class FhirResourceDeserializer(DefaultFieldDeserializer):
             return IFhirResource(self.field).fromUnicode(value)
         elif isinstance(value, dict):
             return IFhirResource(self.field).from_dict(value)
+        elif isinstance(value, BaseModel):
+            return IFhirResource(self.field).from_resource_model(value)
+        elif IFhirResourceValue.providedBy(value):
+            return value
         else:
             raise ValueError(
                 "Invalid data type({0}) provided! only dict or "